+++ resources/views/pages/partials/blocks/rich-text.blade.php
@@ -1,11 +1,5 @@
-<div class="wb-stack wb-gap-2">
-    <p class="wb-m-0">{!! nl2br(e($block->content)) !!}</p>
-</div>
+@php($content = trim((string) ($block->content ?? '')))
 
-@if ($block->children->isNotEmpty())
-    <div class="wb-stack wb-gap-4">
-        @foreach ($block->children as $child)
-            @include('pages.partials.block', ['block' => $child])
-        @endforeach
-    </div>
+@if ($content !== '')
+    <div class="wb-stack wb-gap-2">{!! $content !!}</div>
 @endif